Order files by their path
On macOS (unlike on Windows), `Glob.ExpandNames` does not sort the files by their path. For some reason that I have not investigated, some tests related to _IS24_ fail when the files are not generated in the expected order.

In order to reproduce this issue, one can change the introduced `OrderBy` with `OrderByDescending`.