Fix invalid interface generated for attribute group
Fixes some instances when a invalid interface was generated for
attribute groups containing a attribute with a `list` type.

Correct logic was already implemented in the function used to
generate interfaces for "normal" groups. This logic was moved
to a property so that it can be used for both attribute groups
and "normal" groups.

fixes #325